add subscription page

+++ cmd/web/render.go
@@ -29,8 +29,8 @@ var function = template.FuncMap{
 }
 
 func formatCurrency(n int) string {
-	f := float32(n / 100)
-	// f := float32(n) / float32(100)
+	// f := float32(n / 100)
+	f := float32(n) / float32(100)
 	return fmt.Sprintf("$%.2f", f)
 }
